Moved custom_course to account modal
