+++ webpack.config.js
@@ -22,7 +22,6 @@ const pages = [
   'contact',
   'whatisnew',
   'contribute',
-  'custom_course',
   'admin',
 ];
